+++ packages/terra-application-header-layout/CHANGELOG.md
@@ -3,4 +3,4 @@ Changelog
 
 Unreleased
 ----------
-* Initial release of terra-application-toolbar
+* Initial release of terra-application-header-layout